msm: mdss: Add support for LPAE systems
In LPAE systems physical memory addressing is different than CPU
addressing scheme. This change enables MDSS driver for LPAE systems
where driver is directly dealing with physical addresses.

+++ drivers/video/msm/mdss/mdss_io_util.c
@@ -101,7 +101,7 @@ int msm_dss_ioremap_byname(struct platform_device *pdev,
 		return -ENODEV;
 	}
 
-	io_data->len = resource_size(res);
+	io_data->len = (u32)resource_size(res);
 	io_data->base = ioremap(res->start, io_data->len);
 	if (!io_data->base) {
 		DEV_ERR("%pS->%s: '%s' ioremap failed\n",
